Rename all category files
