+++ curriculum-ui/src/action-mappers/getall-categories-action-mappers.ts
@@ -1,5 +1,5 @@
 import { Dispatch } from "redux";
-import { FetchAllCategories } from '../remote/Category'
+import { FetchAllCategories } from '../remote/category-remote'
 
 
 export const categoriesTypes = {
