+++ curriculum-ui/src/components/create-skill-component/CreateSkillContainer.ts
@@ -1,6 +1,6 @@
 import { IState } from "../../reducers"
 import { createSkillActionMapper } from "../../action-mappers/skill-action-mapper"
-import { getAllCategoriesActionMapper } from "../../action-mappers/categories-getall-action-mappers"
+import { getAllCategoriesActionMapper } from "../../action-mappers/getall-categories-action-mappers"
 import { connect } from "react-redux"
 import { CreateSkillComponent } from "./CreateSkillComponent"
 const mapStateToProps = (state:IState) => {
