+++ curriculum-ui/src/components/view-all-categories-component/ViewAllCategoriesContainer.ts
@@ -1,12 +1,11 @@
 import { IState } from "../../reducers"
 import { ViewAllCategoriesComponent } from "./ViewAllCategoriesComponent"
-import { getAllCategoriesActionMapper } from "../../action-mappers/categories-getall-action-mappers"
+import { getAllCategoriesActionMapper } from "../../action-mappers/getall-categories-action-mappers"
 import { connect } from "react-redux"
 
 const mapStateToProps = (state:IState) => {
     return {
         allCategory: state.allCategory.allCategory,
-        //loggedUser: state.loggedUser.loggedUser,
         errorMessage: state.allCategory.errorMessage        
     }
 }
